// packages/pack/content/server-entity/interfaces/behaviors/look-at-target-behavior.ts
import { BehaviorPriority } from './behavior-priority';

/**
 * Allows the mob to look at the entity they are targetting.
 */
export interface LookAtTargetBehavior extends BehaviorPriority {
  /**
   * The distance in blocks from which the entity will look at this mob's current target
   * @default 8.0
   */
  lookDistance?: number;

  /**
   * The probability of looking at the target. A value of 1.00 is 100%
   * @default 0.02
   */
  probability?: number;

  /**
   * Time range to look at this mob's current target
   * @default [2, 4]
   */
  lookTime?: [number, number];

  /**
   * The angle in degrees that the mob can see in the X-axis (left-right)
   * @default 360
   */
  angleOfViewVertical?: number;

  /**
   * The angle in degrees that the mob can see in the Y-axis (up-down)
   * @default 360
   */
  angleOfViewHorizontal?: number;
}